[#123] refactor : security exception 파일 위치 변경
