+++ src/main/java/com/team/delightserver/security/exception/InvalidOAuth2AttributesException.java
@@ -1,4 +1,4 @@
-package com.team.delightserver.security.oauth2.exception;
+package com.team.delightserver.security.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
